backend/requirements.txt
# Django framework
Django>=4.2,<5.0

# Django REST Framework for building APIs
djangorestframework>=3.14.0,<4.0

# JWT authentication for DRF
djangorestframework-simplejwt>=5.2.2,<6.0

# CORS headers to allow cross-origin requests from Next.js frontend
django-cors-headers>=3.15.0,<4.0

# PostgreSQL adapter for Python
psycopg2-binary>=2.9.6,<3.0
